Rename logging_db tests directory to display
We are planning to put the tests for the actual logging db under a tests
folder called logging_db, but we were already using that name for the
functionality inside of display. Therefore we renamed the display tests
to display, which matches the names of the respective modules in
chalk-solve.

Co-authored-by: David Ross <daboross@daboross.net>