Add test harness for LoggingRustIrDatabase
The harness takes in a program and a goal using the same macro interface
as the existing tests. It will then solve the goal, output the logged
program, and then re-solve the goal to ensure all the information needed
to solve the goal was captured.

Includes a test we used to verify the harness worked.

This also changes a few things in the existing test harness, mainly
splitting `test!()` into two macros - the other being
`parse_test_data!()`, so we can reuse the parsing without calling
solve_goal afterwards.

Co-authored-by: David Ross <daboross@daboross.net>