Add test harness for LoggingRustIrDatabase
The harness takes in a program and a goal using the same macro interface
as the existing tests. It will then solve the goal, output the logged
program, and then re-solve the goal to ensure all the information needed
to solve the goal was captured.

Includes a test we used to verify the harness worked.

Co-authored-by: David Ross <daboross@daboross.net>