Implement stub items in logging_db
This implements the suggested fix where if LoggingRustIrDatabase finds
an name in the output which it isn't already outputting, it'll prepend
a "stub" version of that thing - it'll have the right generics, but no
where bounds and no content. Structs are empty, traits have no methods,
etc.

This doesn't entirely work - mainly, there are some bugs with
associated type bounds not being visited to find missing IDs, and we
don't have enough tests - but the framework is there!