Fix logging db missing assoc ty value names
The logging db that wraps RustIrDatabase was not collecting the ids for
items referenced in assocated type values.

The visitor we are repurposing from the solver does not cross identifier
boundaries. The `AssocTyDatum` stores the values as a set of ids, so
they aren't visited.

We extended the id collector to resolve the assoc ty value ids in an
impl block, and then visit them.

Co-authored-by: David Ross <daboross@daboross.net>