Add Id Collector unit tests
Currently, these don't work due to a circular depdency that causes
bounds checks to fail incorrectly.

Specifically, chalk-integration depends on the non-test version of
chalk-solve, while chalk-solve test depends on the non-test version of
chalk-integration. This means we end up with both a test and a non-test
version of chalk-solve. This resulted in a duplicated ChalkIrDatabase
trait, where one version was unable to be used to satisfy the other
version.