Relate at top level during generalization
This fixed an issue where we failed to create constraints of lifetimes
in the dyn_lifetime_bound and other tests. This was because we weren't
instantiating the new lifetime variables universally and existentially.

The fix was to call relate_ty_ty on the entire structure, as opposed to
only calling it on the newly generalized inference variables. Traversing
from the top uses the existing code to add the required constraints.

Co-authored-by: David Ross <daboross@daboross.net>